Added SDO filter to set torque mode function
Decided that all configuration functions should be selfish and ifgnore all other incoming CAN messages that are not responses to the messages they send out. This is not good practice in general, but for our purposes, this is ok. When the startup procedure is happening, each step is pre-defined and no other data should be important between steps in the procedure. So I made the torque configuration function filter out the SDO confirmation message form any junk that may be leftover in the CAN read buffer. I still need to incorporate wether the operating mode was changed or not in the functions returned variable.

+++ CAN_Test_Code/ProgreSSIV_MC_Driver.h
@@ -16,10 +16,10 @@
 
 #define SCALE_FACTOR 2
 
-int  reset_nodes();
-int  reset_communications();
+uint8_t  reset_nodes();
+uint8_t  reset_communications();
 int  start_remote_nodes();
-int  set_torque_operating_mode(int node_id);
+uint8_t  set_torque_operating_mode(int node_id);
 
 int controlword_operation_enabled(int node_id);
